refactor: 파일 통합, getBroswer 로직 수정
